Leap Year Detection
Leap year is like a normal year except one additional day at the end of Februrary. Normally, There are 365 days in a year. But, In Leap Year, Feburary gets an additional day at the end of the month. We can detect if it is Leap year or not. For that, There are some conditons to must be checked:
1. On every year that is divisible by 4 with no remainder
2. Except every year that is evenly divisible by 100 with no remainder
3. Unless the year is also divisible by 400 with no remainder