Caesar Cipher Encryption
The Caesar Cipher is one of the oldest encryption methods. The idea involves comparing the normal alphabet with a shifted alphabet determined by a specified number, and selecting the letters in the same column. Not that hard to be decrypt in 21th century, However, In that times, This method hid a lot of truth from curious eyes.